@@ -1,12 +1,9 @@
 package id.odt.museek.fragment;
 
-import android.content.Context;
 import android.content.Intent;
-import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -24,10 +21,8 @@ import butterknife.ButterKnife;
 import de.hdodenhof.circleimageview.CircleImageView;
 import id.odt.museek.R;
 import id.odt.museek.activity.AddMusicActivity;
-import id.odt.museek.activity.HomeActivity;
 import id.odt.museek.activity.LoginActivity;
 import id.odt.museek.activity.SettingActivity;
-import id.odt.swipecardlib.SwipeCardView;
 
 public class ProfileFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
