@@ -3,12 +3,11 @@ package id.odt.museek.ui;
 import android.content.Context;
 import android.graphics.Typeface;
 import android.util.AttributeSet;
-import android.widget.EditText;
 
 /**
  * Created by arrival on 3/12/15.
  */
-public class MyEditText extends android.support.v7.widget.AppCompatEditText {
+public class MyEditText extends androidx.appcompat.widget.AppCompatEditText {
 
     public MyEditText(Context context, AttributeSet attrs, int defStyle) {
         super(context, attrs, defStyle);
